refactor: 주어진 클래스 적절한 위치로 이동
