+++ src/main/java/bridge/controller/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.controller;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
