+++ src/main/java/bridge/view/output/OutputView.java
@@ -1,9 +1,15 @@
-package bridge;
+package bridge.view.output;
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
 public class OutputView {
+
+    private final Printer printer;
+
+    public OutputView(Printer printer) {
+        this.printer = printer;
+    }
 
     /**
      * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
